remove warnings

+++ Scrabble_BackEnd/BStarNode.cpp
@@ -2,7 +2,7 @@
 
 
 
-BStarNode::BStarNode(int opt, int pess, int id,const Move& move)
+BStarNode::BStarNode(double opt, double pess, int id,const Move& move)
 {
 	this->optm = opt;
 	this->pess = pess;
